Moved test utilities to MDAnalysisTests/util
Removed check for assert_ import, should be covered in setup.py rather
than runtime
