+++ testsuite/MDAnalysisTests/__init__.py
@@ -128,19 +128,10 @@ except ImportError:
 #  code won't be run again under coverage's watch. See Issue 344.
 
 import os
-from contextlib import contextmanager
+import sys
 
 # We get our nose from the plugins so that version-checking needs only be done there.
 from MDAnalysisTests.plugins import nose, loaded_plugins
-import sys
-import importlib
-# This is a de facto test for numpy's version, since we don't actually need assert_ here.
-#  Should we be clean about this and just call distutils to compare version strings?
-try:
-    from numpy.testing import assert_
-except ImportError:
-    raise ImportError("""numpy>=1.5  is required to run the test suite. Please install it first. """
-                      """(For example, try "easy_install 'numpy>=1.5'").""")
 
 # Any tests that plot with matplotlib need to run with the simple agg backend because
 # on Travis there is no DISPLAY set
@@ -150,7 +141,13 @@ try:
 except ImportError:
     pass
 
-from MDAnalysisTests.util import block_import
+from MDAnalysisTests.util import (
+    block_import,
+    executable_not_found,
+    module_not_found,
+    parser_not_found,
+    in_dir,
+)
 
 def run(*args, **kwargs):
     """Test-running function that loads plugins, sets up arguments, and calls `nose.run_exit()`"""
@@ -175,77 +172,3 @@ def run(*args, **kwargs):
     return nose.run_exit(*args, **kwargs)
 
 
-def executable_not_found(*args):
-    """Return ``True`` if none of the executables in args can be found.
-
-    ``False`` otherwise (i.e. at least one was found).
-
-    To be used as the argument of::
-
-    @dec.skipif(executable_not_found("binary_name"), msg="skip test because binary_name not available")
-    """
-    # This must come here so that MDAnalysis isn't imported prematurely,
-    #  which spoils coverage accounting (see Issue 344).
-    import MDAnalysis.lib.util
-    for name in args:
-        if MDAnalysis.lib.util.which(name) is not None:
-            return False
-    return True
-
-
-def module_not_found(module):
-    try:
-        importlib.import_module(module)
-    except ImportError:
-        return True
-    else:
-        return False
-
-
-def parser_not_found(parser_name):
-    """Return ``True`` if the parser of the given name cannot be found.
-
-    This allows to skip a test when the parser is unavailable (e.g in python 3
-    when the parser is not compatible).
-
-    To be used as the argument of::
-
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser is not available. Are you on python 3?')
-    """
-    import MDAnalysis.coordinates
-    try:
-        getattr(MDAnalysis.coordinates, parser_name)
-    except AttributeError:
-        return True
-    else:
-        return False
-
-@contextmanager
-def in_dir(dirname):
-    """Context manager for safely changing directories.
-
-    Arguments
-    ---------
-    dirname : string
-        directory to change into
-
-    Example
-    -------
-    Change into a temporary directory and always change back to the
-    current one::
-
-      with in_dir("/tmp") as tmpdir:
-          # do stuff
-
-    SeeAlso
-    -------
-    The :mod:`tmpdir` module provides functionality such as :func:`tmpdir.in_tmpdir`
-    to create temporary directories that are automatically deleted once they are no
-    longer used.
-    """
-
-    old_path = os.getcwd()
-    os.chdir(dirname)
-    yield dirname
-    os.chdir(old_path)
