+++ testsuite/MDAnalysisTests/util.py
@@ -33,8 +33,11 @@ except ImportError:
     builtins_name = 'builtins'
     importer = builtins.__import__
 
+from contextlib import contextmanager
 from functools import wraps
+import importlib
 import mock
+import os
 
 
 def block_import(package):
@@ -66,3 +69,78 @@ def block_import(package):
     return blocker_wrapper
 
 
+def executable_not_found(*args):
+    """Return ``True`` if none of the executables in args can be found.
+
+    ``False`` otherwise (i.e. at least one was found).
+
+    To be used as the argument of::
+
+    @dec.skipif(executable_not_found("binary_name"), msg="skip test because binary_name not available")
+    """
+    # This must come here so that MDAnalysis isn't imported prematurely,
+    #  which spoils coverage accounting (see Issue 344).
+    import MDAnalysis.lib.util
+    for name in args:
+        if MDAnalysis.lib.util.which(name) is not None:
+            return False
+    return True
+
+
+def module_not_found(module):
+    try:
+        importlib.import_module(module)
+    except ImportError:
+        return True
+    else:
+        return False
+
+
+def parser_not_found(parser_name):
+    """Return ``True`` if the parser of the given name cannot be found.
+
+    This allows to skip a test when the parser is unavailable (e.g in python 3
+    when the parser is not compatible).
+
+    To be used as the argument of::
+
+    @dec.skipif(parser_not_found('DCD'),
+                'DCD parser is not available. Are you on python 3?')
+    """
+    import MDAnalysis.coordinates
+    try:
+        getattr(MDAnalysis.coordinates, parser_name)
+    except AttributeError:
+        return True
+    else:
+        return False
+
+
+@contextmanager
+def in_dir(dirname):
+    """Context manager for safely changing directories.
+
+    Arguments
+    ---------
+    dirname : string
+        directory to change into
+
+    Example
+    -------
+    Change into a temporary directory and always change back to the
+    current one::
+
+      with in_dir("/tmp") as tmpdir:
+          # do stuff
+
+    SeeAlso
+    -------
+    The :mod:`tmpdir` module provides functionality such as :func:`tmpdir.in_tmpdir`
+    to create temporary directories that are automatically deleted once they are no
+    longer used.
+    """
+
+    old_path = os.getcwd()
+    os.chdir(dirname)
+    yield dirname
+    os.chdir(old_path)
